+++ examples/Demo/Source/Demos/OpenGLDemo2D.cpp
@@ -74,7 +74,7 @@ public:
 
     void paint (Graphics& g) override
     {
-        g.fillCheckerBoard (getLocalBounds(), 48, 48, Colours::lightgrey, Colours::white);
+        g.fillCheckerBoard (getLocalBounds().toFloat(), 48.0f, 48.0f, Colours::lightgrey, Colours::white);
 
         if (shader == nullptr || shader->getFragmentShaderCode() != fragmentCode)
         {
